Refactor Complete ✔️, project now has two modules!! Much faster compiling and much much leaner runtime! None of the MCP scripts will be compiled into builds.
