+++ GenerativeAISupport.uplugin
@@ -19,6 +19,11 @@
 			"Name": "GenerativeAISupport",
 			"Type": "Runtime",
 			"LoadingPhase": "Default"
+		},
+		{
+			"Name": "GenerativeAISupportEditor",
+			"Type": "Editor",
+			"LoadingPhase": "PostEngineInit"
 		}
 	]
 }
