+++ Source/GenerativeAISupport/GenerativeAISupport.Build.cs
@@ -1,4 +1,8 @@
-// Copyright Epic Games, Inc. All Rights Reserved.
+// Copyright (c) 2025 Prajwal Shetty. All rights reserved.
+// Licensed under the MIT License. See LICENSE file in the root directory of this
+// source tree or http://opensource.org/licenses/MIT.
+
+
 
 using UnrealBuildTool;
 
@@ -8,79 +12,24 @@ public class GenerativeAISupport : ModuleRules
 	{
 		PCHUsage = ModuleRules.PCHUsageMode.UseExplicitOrSharedPCHs;
 
-		PublicIncludePaths.AddRange(
-			new string[]
-			{
-				// ... add public include paths required here ...
-			}
-		);
-
-
-		PrivateIncludePaths.AddRange(
-			new string[]
-			{
-				// ... add other private include paths required here ...
-			}
-		);
-
-
 		PublicDependencyModuleNames.AddRange(
 			new string[]
 			{
-				"Core", "Engine",
-				"ImageDownload",
-				"ImageDownload",
+				"Core",
+				"CoreUObject",
+				"Engine",
+				"HTTP",
+				"Json",
 				"DeveloperSettings",
-				"AutomationTest",
-				"FunctionalTesting", 
-				"EditorStyle"
-				// ... add other public dependencies that you statically link with here ...
+				"ImageDownload"
 			}
 		);
-
 
 		PrivateDependencyModuleNames.AddRange(
 			new string[]
 			{
-				"CoreUObject",
-				"Engine",
 				"Slate",
-				"SlateCore",
-				"Json",
-				"HTTP",
-				"EditorScriptingUtilities",
-				"Blutility",
-				"UnrealEd",
-				"MaterialEditor",
-				"MaterialUtilities",
-				"BlueprintGraph",
-				"FunctionalTesting", 
-				"UMGEditor",
-				"UMG", 
-				"EditorStyle", 
-				"WorkspaceMenuStructure",
-				"Projects"
-				// ... add private dependencies that you statically link with here ...	
-			}
-		);
-
-		if (Target.Type == TargetRules.TargetType.Editor)
-		{
-			PrivateDependencyModuleNames.AddRange(
-				new string[]
-				{
-					// Your existing dependencies...
-					"Settings"
-					// PythonScriptPlugin is no longer needed here
-				}
-			);
-		}
-
-
-		DynamicallyLoadedModuleNames.AddRange(
-			new string[]
-			{
-				// ... add any modules that your module loads dynamically here ...
+				"SlateCore"
 			}
 		);
 	}
