+++ Source/GenerativeAISupport/Private/Secure/GenSecureKey.cpp
@@ -1,4 +1,7 @@
-// Copyright Prajwal Shetty 2024. All rights Reserved. https://prajwalshetty.com/terms
+// Copyright (c) 2025 Prajwal Shetty. All rights reserved.
+// Licensed under the MIT License. See LICENSE file in the root directory of this
+// source tree or http://opensource.org/licenses/MIT.
+
 
 #include "Secure/GenSecureKey.h"
 #include "Data/GenAIOrgs.h"
