+++ Source/GenerativeAISupport/Public/GenerativeAISupport.h
@@ -13,17 +13,4 @@ public:
     // IModuleInterface implementation
     virtual void StartupModule() override;
     virtual void ShutdownModule() override;
-
-private:
-    void RegisterSettings();
-    void UnregisterSettings();
-    
-    bool HandleSettingsSaved();
-    void RegisterMenuExtension();
-    void UnregisterMenuExtension();
-    void OnEditorWindowMenuClicked();
-
-    bool bSettingsRegistered;
-    
-    TSharedPtr<FUICommandList> PluginCommands;
 };
