+++ Source/GenerativeAISupportEditor/Private/GenerativeAISupportSettings.cpp
@@ -1,4 +1,7 @@
-// Copyright Epic Games, Inc. All Rights Reserved.
+// Copyright (c) 2025 Prajwal Shetty. All rights reserved.
+// Licensed under the MIT License. See LICENSE file in the root directory of this
+// source tree or http://opensource.org/licenses/MIT.
+
 
 #include "GenerativeAISupportSettings.h"
