+++ Source/GenerativeAISupportEditor/Public/GenerativeAISupportSettings.h
@@ -1,4 +1,7 @@
-// Copyright Epic Games, Inc. All Rights Reserved.
+// Copyright (c) 2025 Prajwal Shetty. All rights reserved.
+// Licensed under the MIT License. See LICENSE file in the root directory of this
+// source tree or http://opensource.org/licenses/MIT.
+
 
 #pragma once
 
@@ -10,7 +13,7 @@
  * Settings for the Generative AI Support Plugin
  */
 UCLASS(config=Game, defaultconfig)
-class GENERATIVEAISUPPORT_API UGenerativeAISupportSettings : public UObject
+class GENERATIVEAISUPPORTEDITOR_API  UGenerativeAISupportSettings : public UObject
 {
     GENERATED_BODY()
